Refactor UI as CompositeController
